agora-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sse-starlette==2.1.0
aiosqlite==0.20.0
pydantic[email]==2.7.0
python-dotenv==1.0.1
langchain==0.2.3
langchain-groq==0.1.5
langchain-community==0.2.3
langgraph==0.1.5
tavily-python==0.3.3
faiss-cpu==1.8.0
beautifulsoup4==4.12.3
requests==2.31.0
httpx==0.27.0
python-multipart==0.0.9
websockets==12.0
